[refactor] Moved the candidate tables into its own component (#74)
